Begin refactoring and use separate files for commands
