+++ package.json
@@ -1,5 +1,6 @@
 {
   "dependencies": {
+    "@discordjs/rest": "^0.3.0",
     "axios": "^0.21.1",
     "discord.js": "^13.2.0",
     "dotenv": "^8.2.0",
