--- slashCommands.json
@@ -1,137 +1,0 @@
-[
-  {
-    "name": "train",
-    "description": "Sends a single training question to be answered",
-    "options": [
-      {
-        "type": 3,
-        "name": "subject",
-        "description": "Optional subject to be used as a filter",
-        "default": false,
-        "required": false,
-        "choices": [
-          {
-            "name": "astro",
-            "value": "astro"
-          },
-          {
-            "name": "bio",
-            "value": "bio"
-          },
-          {
-            "name": "ess",
-            "value": "ess"
-          },
-          {
-            "name": "chem",
-            "value": "chem"
-          },
-          {
-            "name": "phys",
-            "value": "phys"
-          },
-          {
-            "name": "math",
-            "value": "math"
-          },
-          {
-            "name": "energy",
-            "value": "energy"
-          }
-        ]
-      }
-    ]
-  },
-  {
-    "name": "help",
-    "description": "Replies with a help message explaining what the bot can do"
-  },
-  {
-    "name": "rounds",
-    "options": [
-      {
-        "type": 1,
-        "name": "generate",
-        "description": "Generates a round with randomized questions from https://scibowldb.com/",
-        "options": []
-      },
-      {
-        "type": 1,
-        "name": "list",
-        "description": "Lists your 5 most recently generated rounds with links",
-        "options": []
-      },
-      {
-        "type": 1,
-        "name": "hit",
-        "description": "Shows the total number of rounds hit as well as the number for the specific user",
-        "options": []
-      }
-    ],
-    "description": "Commands regarding rounds generated by AwesomeSciBo"
-  },
-  {
-    "name": "top",
-    "description": "Lists top ten scores across servers (server specific leaderboard WIP)"
-  },
-  {
-    "name": "about",
-    "options": [
-      {
-        "type": 1,
-        "name": "contributors",
-        "description": "Lists contributors to the AwesomeSciBo bot",
-        "options": []
-      },
-      {
-        "type": 1,
-        "name": "changelog",
-        "description": "Lists the 5 most recent changes in a \"git log\" type format",
-        "options": []
-      },
-      {
-        "type": 1,
-        "name": "bot",
-        "description": "Lists information about AwesomeSciBo",
-        "options": []
-      }
-    ],
-    "description": "Commands regarding the creation/development of the bot"
-  },
-  {
-    "name": "result",
-    "description": "The result of a regional",
-    "options": [
-      {
-        "type": "STRING",
-        "name": "location",
-        "description": "The location of the regional",
-        "required": true
-      },
-      {
-        "type": "STRING",
-        "name": "winner",
-        "description": "The advancing team from the regional",
-        "required": true
-      },
-      {
-        "type": "STRING",
-        "name": "runner-up",
-        "description": "The second-place team from the regional",
-        "required": false
-      },
-      {
-        "type": "STRING",
-        "name": "third-place",
-        "description": "The third-place team from the regional",
-        "required": false
-      },
-      {
-        "type": "STRING",
-        "name": "honorable-mentions",
-        "description": "Any extra teams to be mentioned or a space for technicalities to be explained",
-        "required": false
-      }
-    ]
-  }
-]
